走った距離取得処理を簡略化

+++ app/src/main/java/com/reo/running/runnershigh/RecordDatabase.kt
@@ -17,7 +17,7 @@ data class JustRunData(
     @PrimaryKey(autoGenerate = true)
     val id: Int,
     val time: String,
-    val distance: Double,
+    val distance: Float,
     val calorie: Int,
     val runDate: String,
     val bitmap: Bitmap?,
@@ -30,7 +30,7 @@ data class RunResult(
     val id: Int,
     val bitmap: Bitmap?,
     val time: String,
-    val distance: Double,
+    val distance: Float,
     val calorie: Int,
     val runData: String,
     val colorId: String,
